requirements.txt
python-nmap
netifaces
paramiko
requests
pysmb